# watchdog/openapi.yaml
openapi: 3.1.0
info:
  title: Watchdog with Your Infrastructure
  description: Unified API for managing Tempo, Loki, AlertManager, Grafana, and Resolver
  version: 0.0.1
paths:
...
  /api/resolver/analyze/config-template:
    get:
      tags:
      - resolver
      summary: Get Analyze Config Template
      operationId: get_analyze_config_template_api_resolver_analyze_config_template_get
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeConfigTemplateResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '502':
          description: Bad Gateway
      security:
      - HTTPBearer: []
  /api/resolver/analyze/jobs:
    post:
      tags:
      - resolver
      summary: Create Analyze Job
      operationId: create_analyze_job_api_resolver_analyze_jobs_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequestPayload'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeJobCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
    get:
      tags:
      - resolver
      summary: List Analyze Jobs
      operationId: list_analyze_jobs_api_resolver_analyze_jobs_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/AnalyzeJobStatus'
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeJobListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '502':
          description: Bad Gateway
  /api/resolver/analyze/jobs/{job_id}:
    get:
      tags:
      - resolver
      summary: Get Analyze Job
      operationId: get_analyze_job_api_resolver_analyze_jobs__job_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Job Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeJobSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '404':
          description: Not Found
        '502':
          description: Bad Gateway
  /api/resolver/analyze/jobs/{job_id}/result:
    get:
      tags:
      - resolver
      summary: Get Analyze Job Result
      operationId: get_analyze_job_result_api_resolver_analyze_jobs__job_id__result_get
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Job Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeJobResultResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '404':
          description: Not Found
        '502':
          description: Bad Gateway
  /api/resolver/reports/{report_id}:
    get:
      tags:
      - resolver
      summary: Get Report By Id
      operationId: get_report_by_id_api_resolver_reports__report_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Report Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeReportResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '404':
          description: Not Found
        '502':
          description: Bad Gateway
    delete:
      tags:
      - resolver
      summary: Delete Report By Id
      operationId: delete_report_by_id_api_resolver_reports__report_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[A-Za-z0-9_-]+$
          title: Report Id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeReportDeleteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '502':
          description: Bad Gateway
...
  /api/resolver/anomalies/logs/patterns:
    post:
      tags:
      - resolver
      summary: Anomalies Log Patterns
      operationId: anomalies_log_patterns_api_resolver_anomalies_logs_patterns_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeProxyPayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
                type: object
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
      security:
      - HTTPBearer: []
  /api/resolver/anomalies/logs/bursts:
    post:
      tags:
      - resolver
      summary: Anomalies Log Bursts
      operationId: anomalies_log_bursts_api_resolver_anomalies_logs_bursts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeProxyPayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
                type: object
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
      security:
      - HTTPBearer: []
  /api/resolver/anomalies/traces:
    post:
      tags:
      - resolver
      summary: Anomalies Traces
      operationId: anomalies_traces_api_resolver_anomalies_traces_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeProxyPayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
                type: object
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
      security:
      - HTTPBearer: []
  /api/resolver/correlate:
    post:
      tags:
      - resolver
      summary: Correlate Signals
      operationId: correlate_signals_api_resolver_correlate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeProxyPayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
                type: object
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
      security:
      - HTTPBearer: []
  /api/resolver/topology/blast-radius:
    post:
      tags:
      - resolver
      summary: Topology Blast Radius
      operationId: topology_blast_radius_api_resolver_topology_blast_radius_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeProxyPayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
                type: object
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
      security:
      - HTTPBearer: []
  /api/resolver/slo/burn:
    post:
      tags:
      - resolver
      summary: Slo Burn
      operationId: slo_burn_api_resolver_slo_burn_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeProxyPayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
                type: object
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
      security:
      - HTTPBearer: []
  /api/resolver/forecast/trajectory:
    post:
      tags:
      - resolver
      summary: Forecast Trajectory
      operationId: forecast_trajectory_api_resolver_forecast_trajectory_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeProxyPayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
                type: object
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
      security:
      - HTTPBearer: []
  /api/resolver/causal/granger:
    post:
      tags:
      - resolver
      summary: Causal Granger
      operationId: causal_granger_api_resolver_causal_granger_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeProxyPayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
                type: object
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
      security:
      - HTTPBearer: []
  /api/resolver/causal/bayesian:
    post:
      tags:
      - resolver
      summary: Causal Bayesian
      operationId: causal_bayesian_api_resolver_causal_bayesian_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeProxyPayload'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/JSONValue'
                type: object
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '502':
          description: Bad Gateway
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    AnalyzeConfigTemplateResponse:
      properties:
        version:
          type: integer
          title: Version
        defaults:
          additionalProperties:
            $ref: '#/components/schemas/JSONValue'
          type: object
          title: Defaults
        template_yaml:
          type: string
          title: Template Yaml
        file_name:
          type: string
          title: File Name
      type: object
      required:
      - version
      - defaults
      - template_yaml
      - file_name
      title: AnalyzeConfigTemplateResponse
    AnalyzeJobCreateResponse:
      properties:
        job_id:
          type: string
          title: Job Id
        report_id:
          type: string
          title: Report Id
        status:
          $ref: '#/components/schemas/AnalyzeJobStatus'
        created_at:
          type: string
          format: date-time
          title: Created At
        tenant_id:
          type: string
          title: Tenant Id
        requested_by:
          type: string
          title: Requested By
      type: object
      required:
      - job_id
      - report_id
      - status
      - created_at
      - tenant_id
      - requested_by
      title: AnalyzeJobCreateResponse
...
    JSONValue:
      anyOf: *id002
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    WatchdogCookieAuth:
      type: apiKey
      in: cookie
      name: watchdog_token
      description: Authenticate using the watchdog_token cookie (same token as Bearer).
...